Smart Attendance System/code/requirements.txt
pyzbar == 0.1.9
pillow == 10.2.0
opencv-contrib-python == 4.9.0.80
opencv-python == 4.9.0.80
